+++ PixivFSUWP/MainPage.xaml
@@ -48,13 +48,6 @@
                         <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" 
                                   FontSize="16" Glyph="&#xE721;"/>
                     </AppBarButton>
-                    <AppBarButton Width="48" x:Name="btnGoToID" Click="BtnGoTo_Click">
-                        <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" 
-                                  FontSize="16" Glyph="&#xE751;"/>
-                    </AppBarButton>
-                    <AppBarButton Width="48" x:Name="btnSauceNAO" Click="btnSauceNAO_Click">
-                        <Image Source="/Assets/SauceNAO.png"/>
-                    </AppBarButton>
                     <AppBarButton Width="48" x:Name="btnSetting" Click="BtnSetting_Click">
                         <FontIcon FontFamily="{StaticResource SymbolThemeFontFamily}" 
                                   FontSize="16" Glyph="&#xE713;"/>
