+++ PixivFSUWP/SearchPage.xaml.cs
@@ -18,6 +18,8 @@ using PixivCS;
 using Windows.UI.Xaml.Media.Imaging;
 using PixivFSUWP.Interfaces;
 using static PixivFSUWP.Data.OverAll;
+// 额外内容
+using PixivFSUWP.SauceNao;
 
 // https://go.microsoft.com/fwlink/?LinkId=234238 上介绍了“空白页”项模板
 
@@ -187,6 +189,94 @@ namespace PixivFSUWP
             cbSort.SelectedIndex = 0;
             cbDuration.SelectedIndex = 0;
             TxtWord_QuerySubmitted(null, null);
+        }
+
+        private async void btnSauceNAO_Click(object sender, RoutedEventArgs e)
+        {
+            Frame.Navigate(typeof(SauceNAOPage));
+            Windows.Storage.ApplicationDataContainer localSettings = Windows.Storage.ApplicationData.Current.LocalSettings;
+            string SAUCENAO_API_KEY = localSettings.Values["SauceNAOAPI"] as string;//读取设置项
+            string IMGUR_API_KEY = localSettings.Values["ImgurAPI"] as string;
+            if (SAUCENAO_API_KEY == null)
+            {
+                // "未设置SauceNAO API"
+                Frame.Navigate(typeof(SettingsPage));
+                return;
+                // SAUCENAO_API_KEY = "默认API";
+            }
+            else if (SAUCENAO_API_KEY.Length == 0)
+            {
+                // "未设置SauceNAO API"
+                Frame.Navigate(typeof(SettingsPage));
+                return;
+                // SAUCENAO_API_KEY = "默认API";
+            }
+            if (IMGUR_API_KEY == null)
+            {
+                // "未设置Imger API"
+                Frame.Navigate(typeof(SettingsPage));
+                return;
+                // IMGUR_API_KEY = "默认API";
+            }
+            else if (IMGUR_API_KEY.Length == 0)
+            {
+                // "未设置Imger API"
+                Frame.Navigate(typeof(SettingsPage));
+                return;
+                // IMGUR_API_KEY = "默认API";
+            }
+            var picker = new Windows.Storage.Pickers.FileOpenPicker();
+            picker.ViewMode = Windows.Storage.Pickers.PickerViewMode.Thumbnail;
+            picker.SuggestedStartLocation = Windows.Storage.Pickers.PickerLocationId.PicturesLibrary;
+            picker.FileTypeFilter.Add(".jpg");
+            picker.FileTypeFilter.Add(".jpeg");
+            picker.FileTypeFilter.Add(".png");
+            Windows.Storage.StorageFile file = await picker.PickSingleFileAsync();
+            if (file == null)
+            {
+                Frame.GoBack();
+                return;
+            }
+            byte[] IMAGE_PATH = await StorageFileExt.AsByteArray(file);
+
+            Action<byte[]> action = IMAGE_BYTES =>
+            {
+                Console.WriteLine("==== 以图搜源 ====");
+                string image = Imgur.Upload(IMAGE_BYTES, IMGUR_API_KEY);
+                List<Result> results = new SauceNao.SauceNao(SAUCENAO_API_KEY).Request(image, null);
+                // 这里是调试输出查询结果的内容..
+                //results.RemoveAll(result => !result.HasRecognizableSauce());
+                //foreach (Result result in results)
+                //{
+                //    System.Diagnostics.Debug.WriteLine(result.ToString() + "\n");
+                //}
+                System.Diagnostics.Debug.WriteLine("Pixiv ID = " + results[0].Response.SauceId.ToString());
+                Frame.Navigate(typeof(IllustDetailPage), results[0].Response.SauceId);
+            };
+            action.Invoke(IMAGE_PATH);
+        }
+
+        private void BtnGoTo_Click(object sender, RoutedEventArgs e)
+        {
+            Frame.Navigate(typeof(GoToPIDPage));
+        }
+    }
+    static class StorageFileExt
+    {
+        /// <summary>
+        /// 将文件转换为字节数组
+        /// </summary>
+        /// <param name="file"></param>
+        /// <returns></returns>
+        public static async Task<byte[]> AsByteArray(this Windows.Storage.StorageFile file)
+        {
+            Windows.Storage.Streams.IRandomAccessStream fileStream =
+                await file.OpenAsync(Windows.Storage.FileAccessMode.Read);
+            var reader = new Windows.Storage.Streams.DataReader(fileStream.GetInputStreamAt(0));
+            await reader.LoadAsync((uint)fileStream.Size);
+            byte[] pixels = new byte[fileStream.Size];
+            reader.ReadBytes(pixels);
+            return pixels;
         }
     }
 }
